fix natfeats test-code for 16-bit compilers
AHCC sizeof() returns word and probably so does e.g. "gcc -mshort".
Thanks & apologies to Nicolas for needing to debug this!

+++ tests/natfeats/natfeats.c
@@ -109,7 +109,7 @@ static void nf_showname(void)
 		long chars;
 		char buffer[64];
 		id |= 0x0001;  /* name + version */
-		chars = nf_call(id, buffer, sizeof(buffer)-2);
+		chars = nf_call(id, buffer, (long)sizeof(buffer)-2);
 		buffer[chars++] = '\n';
 		buffer[chars++] = '\0';
 		nf_print(buffer);
